+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.11.0] - 2026-03-25
+
 ### Added
 
 - `--file-types` flag to include only specified config file types for validation (inverse of `--exclude-file-types`)